+++ changelog.md
@@ -17,7 +17,7 @@ See also:
 
 ### Added
 
-- There is now a facility to update limited episode data after submission. This is currently limited to preventable factors, pre-existing diabetes status, end date/time of the DKA protocol, and cerebral oedema questions. This may be expanded in the future to allow retrospective audit data collection.
+- There is now a facility to update limited episode data after submission. This is currently limited to preventable factors, pre-existing diabetes status, end date/time of the DKA protocol, and cerebral oedema questions. This may be expanded in the future.
 - A unique patient hash is generated and stored using SHA-256 (x1 client-side, x1 server-side with salt) using patient NHS number and date of birth as input string. This allows episodes relating to the same patient to be linked for audit purposes.
 - The patient postcode is collected and sent with other data to the API where an index of multiple deprivation (IMD) decile is derived and stored. The postcode is not stored.
 - Bicarbonate, glucose and ketones may now optionally be entered and will be printed on the care pathway if provided. Bicarbonate is used for severity scoring as above.